+++ src/19.AlienCodex.sol
@@ -4,10 +4,7 @@ pragma solidity ^0.5.0;
 contract Ownable {
     address private _owner;
 
-    event OwnershipTransferred(
-        address indexed previousOwner,
-        address indexed newOwner
-    );
+    event OwnershipTransferred(address indexed previousOwner, address indexed newOwner);
 
     /**
      * @dev Initializes the contract setting the deployer as the initial owner.
@@ -63,10 +60,7 @@ contract Ownable {
      * @dev Transfers ownership of the contract to a new account (`newOwner`).
      */
     function _transferOwnership(address newOwner) internal {
-        require(
-            newOwner != address(0),
-            "Ownable: new owner is the zero address"
-        );
+        require(newOwner != address(0), "Ownable: new owner is the zero address");
         emit OwnershipTransferred(_owner, newOwner);
         _owner = newOwner;
     }
@@ -93,7 +87,7 @@ contract AlienCodex is Ownable {
         codex.length--;
     }
 
-    function revise(uint i, bytes32 _content) public contacted {
+    function revise(uint256 i, bytes32 _content) public contacted {
         codex[i] = _content;
     }
 }
