+++ src/22.Dex.sol
@@ -8,6 +8,7 @@ import "@openzeppelin/contracts/access/Ownable.sol";
 contract Dex is Ownable {
     address public token1;
     address public token2;
+
     constructor() Ownable(msg.sender) {}
 
     function setTokens(address _token1, address _token2) public onlyOwner {
@@ -15,56 +16,39 @@ contract Dex is Ownable {
         token2 = _token2;
     }
 
-    function addLiquidity(address token_address, uint amount) public onlyOwner {
+    function addLiquidity(address token_address, uint256 amount) public onlyOwner {
         IERC20(token_address).transferFrom(msg.sender, address(this), amount);
     }
 
-    function swap(address from, address to, uint amount) public {
-        require(
-            (from == token1 && to == token2) ||
-                (from == token2 && to == token1),
-            "Invalid tokens"
-        );
-        require(
-            IERC20(from).balanceOf(msg.sender) >= amount,
-            "Not enough to swap"
-        );
-        uint swapAmount = getSwapPrice(from, to, amount);
+    function swap(address from, address to, uint256 amount) public {
+        require((from == token1 && to == token2) || (from == token2 && to == token1), "Invalid tokens");
+        require(IERC20(from).balanceOf(msg.sender) >= amount, "Not enough to swap");
+        uint256 swapAmount = getSwapPrice(from, to, amount);
         IERC20(from).transferFrom(msg.sender, address(this), amount);
         IERC20(to).approve(address(this), swapAmount);
         IERC20(to).transferFrom(address(this), msg.sender, swapAmount);
     }
 
-    function getSwapPrice(
-        address from,
-        address to,
-        uint amount
-    ) public view returns (uint) {
-        return ((amount * IERC20(to).balanceOf(address(this))) /
-            IERC20(from).balanceOf(address(this)));
+    function getSwapPrice(address from, address to, uint256 amount) public view returns (uint256) {
+        return ((amount * IERC20(to).balanceOf(address(this))) / IERC20(from).balanceOf(address(this)));
     }
 
-    function approve(address spender, uint amount) public {
+    function approve(address spender, uint256 amount) public {
         SwappableToken(token1).approve(msg.sender, spender, amount);
         SwappableToken(token2).approve(msg.sender, spender, amount);
     }
 
-    function balanceOf(
-        address token,
-        address account
-    ) public view returns (uint) {
+    function balanceOf(address token, address account) public view returns (uint256) {
         return IERC20(token).balanceOf(account);
     }
 }
 
 contract SwappableToken is ERC20 {
     address private _dex;
-    constructor(
-        address dexInstance,
-        string memory name,
-        string memory symbol,
-        uint256 initialSupply
-    ) ERC20(name, symbol) {
+
+    constructor(address dexInstance, string memory name, string memory symbol, uint256 initialSupply)
+        ERC20(name, symbol)
+    {
         _mint(msg.sender, initialSupply);
         _dex = dexInstance;
     }
