+++ src/23.DexTwo.sol
@@ -8,6 +8,7 @@ import "@openzeppelin/contracts/access/Ownable.sol";
 contract DexTwo is Ownable {
     address public token1;
     address public token2;
+
     constructor() Ownable(msg.sender) {}
 
     function setTokens(address _token1, address _token2) public onlyOwner {
@@ -15,54 +16,38 @@ contract DexTwo is Ownable {
         token2 = _token2;
     }
 
-    function add_liquidity(
-        address token_address,
-        uint amount
-    ) public onlyOwner {
+    function add_liquidity(address token_address, uint256 amount) public onlyOwner {
         IERC20(token_address).transferFrom(msg.sender, address(this), amount);
     }
 
-    function swap(address from, address to, uint amount) public {
-        require(
-            IERC20(from).balanceOf(msg.sender) >= amount,
-            "Not enough to swap"
-        );
-        uint swapAmount = getSwapAmount(from, to, amount);
+    function swap(address from, address to, uint256 amount) public {
+        require(IERC20(from).balanceOf(msg.sender) >= amount, "Not enough to swap");
+        uint256 swapAmount = getSwapAmount(from, to, amount);
         IERC20(from).transferFrom(msg.sender, address(this), amount);
         IERC20(to).approve(address(this), swapAmount);
         IERC20(to).transferFrom(address(this), msg.sender, swapAmount);
     }
 
-    function getSwapAmount(
-        address from,
-        address to,
-        uint amount
-    ) public view returns (uint) {
-        return ((amount * IERC20(to).balanceOf(address(this))) /
-            IERC20(from).balanceOf(address(this)));
+    function getSwapAmount(address from, address to, uint256 amount) public view returns (uint256) {
+        return ((amount * IERC20(to).balanceOf(address(this))) / IERC20(from).balanceOf(address(this)));
     }
 
-    function approve(address spender, uint amount) public {
+    function approve(address spender, uint256 amount) public {
         SwappableTokenTwo(token1).approve(msg.sender, spender, amount);
         SwappableTokenTwo(token2).approve(msg.sender, spender, amount);
     }
 
-    function balanceOf(
-        address token,
-        address account
-    ) public view returns (uint) {
+    function balanceOf(address token, address account) public view returns (uint256) {
         return IERC20(token).balanceOf(account);
     }
 }
 
 contract SwappableTokenTwo is ERC20 {
     address private _dex;
-    constructor(
-        address dexInstance,
-        string memory name,
-        string memory symbol,
-        uint initialSupply
-    ) ERC20(name, symbol) {
+
+    constructor(address dexInstance, string memory name, string memory symbol, uint256 initialSupply)
+        ERC20(name, symbol)
+    {
         _mint(msg.sender, initialSupply);
         _dex = dexInstance;
     }
@@ -90,8 +75,8 @@ contract Drain is SwappableTokenTwo {
         address token1 = instance.token1();
         address token2 = instance.token2();
 
-        uint token1Amount = instance.balanceOf(token1, dex);
-        uint token2Amount = instance.balanceOf(token2, dex);
+        uint256 token1Amount = instance.balanceOf(token1, dex);
+        uint256 token2Amount = instance.balanceOf(token2, dex);
 
         address to = token1Amount > token2Amount ? token1 : token2;
